[ARQ-420] Restore embedded OSGi functionality
