--- containers/osgi-embedded-4.2/src/main/resources/META-INF/services/org.jboss.arquillian.spi.client.container.DeployableContainer
@@ -1,1 +1,0 @@
-org.jboss.arquillian.container.osgi.embedded_4_2.OSGiEmbeddedContainer
