--- containers/osgi-embedded-4.2/src/test/resources/META-INF/services/org.jboss.osgi.spi.framework.OSGiBootstrapProvider
@@ -1,1 +1,0 @@
-org.jboss.osgi.framework.launch.OSGiFrameworkBootstrapProvider
