+++ containers/osgi-embedded/.gitignore
