+++ containers/osgi-embedded/scripts/assembly-bundles.xml
