+++ containers/osgi-embedded/src/main/resources/META-INF/services/org.jboss.arquillian.spi.DeploymentPackager
