+++ containers/osgi-embedded/src/main/resources/META-INF/services/org.jboss.arquillian.spi.client.container.DeployableContainer
@@ -1,0 +1,1 @@
+org.jboss.arquillian.container.osgi.embedded.OSGiEmbeddedContainer
