+++ containers/osgi-embedded/src/test/resources/jboss-osgi-framework.properties
@@ -1,0 +1,16 @@
+#
+# Properties read by the org.jboss.osgi.spi.framework.PropertiesBootstrapProvider
+# 
+
+# Properties to configure the Framework
+org.osgi.framework.storage=./target/osgi-store
+org.osgi.framework.storage.clean=onFirstInit
+
+# Extra System Packages
+org.osgi.framework.system.packages.extra=\
+	org.jboss.logging;version=3.0, \
+  	org.jboss.osgi.deployment.interceptor;version=1.0, \
+  	org.jboss.osgi.spi;version=1.0, \
+  	org.jboss.osgi.spi.capability;version=1.0, \
+  	org.jboss.osgi.spi.util;version=1.0, \
+  	org.jboss.osgi.testing;version=1.0
