+++ containers/pom.xml
@@ -44,13 +44,13 @@
         <module>jbossas-embedded-6</module>
         <module>reloaded-embedded-1</module>
 <!--       Deprecated 
-	<module>jsr88-remote-1.2</module>
+	      <module>jsr88-remote-1.2</module>
         <module>glassfish-remote-3</module>
         <module>glassfish-embedded-3</module> 
-	<module>osgi-embedded-4.2</module>
 -->
-	<module>glassfish-embedded-3.1</module>
-	<module>glassfish-remote-3.1</module>
+	      <module>osgi-embedded</module>
+	      <module>glassfish-embedded-3.1</module>
+	      <module>glassfish-remote-3.1</module>
         <module>jetty-embedded-6.1</module> 
         <module>jetty-embedded-7</module>
         <module>tomcat-embedded-6</module>
