+++ examples/junit/pom.xml
@@ -647,12 +647,6 @@
          <scope>test</scope>
       </dependency>
       <dependency>
-        <groupId>org.jboss.osgi</groupId>
-        <artifactId>jboss-osgi-spi</artifactId>
-         <scope>test</scope>
-      </dependency>
-      
-      <dependency>
          <groupId>junit</groupId>
          <artifactId>junit</artifactId>
          <scope>test</scope>
