+++ protocols/jmx-osgi-bundle/pom.xml
@@ -40,7 +40,7 @@
               javax.xml.parsers,
               org.w3c.dom,
 
-              org.jboss.logging;version="[2.1,3.0)",
+              org.jboss.logging;version="[3.0,4.0)",
 
               org.osgi.framework;version="[1.5,2.0)",
               org.osgi.util.tracker;version="[1.4,2.0)",
@@ -123,7 +123,7 @@
     </dependency>
     <dependency>
       <groupId>org.jboss.logging</groupId>
-      <artifactId>jboss-logging-spi</artifactId>
+      <artifactId>jboss-logging</artifactId>
     </dependency>
     <dependency>
       <groupId>junit</groupId>
