--- protocols/jmx-osgi/src/main/resources/META-INF/services/org.jboss.arquillian.spi.DeploymentPackager
@@ -1,1 +1,0 @@
-org.jboss.arquillian.packager.osgi.OSGiDeploymentPackager
