+++ protocols/jmx-osgi/src/main/resources/META-INF/services/org.jboss.arquillian.spi.client.protocol.Protocol
@@ -1,0 +1,1 @@
+org.jboss.arquillian.protocol.osgi.JMXOSGiProtocol
